+++ lib/async_queue.dart
@@ -7,5 +7,3 @@
 library async_queue;
 
 export 'src/async_queue_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.